Minor refactoring. cleaning up project: velocity templates has more meaningful names. Moved CommitMessageParser logic to CommitManager. Changed variable 'getLink' to more meaningful 'gitlabURL', 'secret'->'gitlabSecretToken'. New: JIRA SectionAction fix. Now, with vm-s, is saving and restoring configuration

+++ src/main/java/pl/hycom/jira/plugins/gitlab/integration/dao/ConfigManagerDao.java
@@ -23,5 +23,5 @@ public interface ConfigManagerDao {
     ConfigEntity getProjectConfig(Long projectID) throws SQLException;
     public List<ConfigEntity> getAllProjectConfigs() throws SQLException;
     ConfigEntity updateProjectConfig(Long projectID,String gitlabLink,String gitlabSecret,String gitlabClientId,String gitlabProjectName) throws SQLException;
-    public void updateGitlabProjectId(Long projectID, int gitlabProjectID);
+    public void updateGitlabProjectId(Long projectID, Long gitlabProjectID);
 }
